backend/requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
joblib>=1.3.0
onnxruntime>=1.18.0
pydantic>=2.0.0
